requirements.txt
streamlit
langchain
langchain-groq
wikipedia
python-dotenv
duckduckgo-search
arxiv